Extract keys and key related utils into their own module
